avoid compiler warning (returning double from a function declared as returning float)
SVN-Revision: 27707

+++ contrib/brl/bseg/boxm/boxm_cell_data_traits.h
@@ -31,9 +31,7 @@ float boxm_cell_to_float(boct_tree_cell<T_loc, bsta_num_obs<bsta_gauss_f1> >* ce
 template <class T_loc, class T_data>
 float boxm_cell_to_float(boct_tree_cell<T_loc, T_data >* cell, double step_len)
 {
-
-  return 1.0 - vcl_exp(-cell->data().basic_val()*step_len);
-
+  return 1.f - float(vcl_exp(-cell->data().basic_val()*step_len));
 }
 
 #endif
